chore: move core/game_animations to animations/game_animations, move setSprite functions to core/sprites
