+++ src/animations/game_animations.c
